% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formats.R
\name{jrc_tech_report}
\alias{jrc_tech_report}
\title{Output formats for the JRC reportss}
\usage{
jrc_tech_report(...)
}
\arguments{
\item{\dots}{Arguments passed to \code{\link[bookdown]{pdf_document2}}.}
}
\value{
An R Markdown output format object.
}
\description{
Each function is a wrapper for \code{\link[bookdown]{pdf_document2}} to
produce documents in JRC style.
}
\seealso{
\code{\link{https://github.com/robjhyndman/MonashEBSTemplates}}
}
\author{
Andrea Blasco
}
